refactor : notion components
