+++ src/components/notion/callout/index.scss
@@ -1,0 +1,19 @@
+.block-callout {
+  line-height: 1.75rem;
+  display: flex;
+  flex-direction: column;
+  background-color: var(--color-shadow);
+  box-shadow: 0px 6px 5px -4px var(--color-grey-3);
+  padding: 12px 16px 12px 4px;
+  padding-left: 40px;
+  margin: 24px 0px;
+  border-radius: 8px;
+  position: relative;
+  color: var(--color-base);
+  .icon-box {
+    position: absolute;
+    top: 10px;
+    left: 10px;
+    font-size: 1.5rem;
+  }
+}
