+++ src/components/notion/heading2/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import '@scss/notion.scss';
+import './index.scss';
 import { TextBlock } from '@types';
 import { NParagraph } from '@components/notion';
