+++ src/components/notion/image/index.scss
@@ -1,0 +1,12 @@
+.block-image {
+  width: 100%;
+  margin: 0;
+  img {
+    width: 100%;
+    border-radius: 4px;
+  }
+  .caption-box {
+    color: var(--color-grey);
+    font-size: 10pt;
+  }
+}
