+++ src/components/notion/index.ts
@@ -1,12 +1,12 @@
-export { default as NHeading1 } from '@components/notion/NHeading1';
-export { default as NHeading2 } from '@components/notion/NHeading2';
-export { default as NHeading3 } from '@components/notion/NHeading3';
-export { default as NBookmark } from '@components/notion/NBookmark';
-export { default as NBulletedList } from '@components/notion/NBulletedList';
-export { default as NCallout } from '@components/notion/NCallout';
-export { default as NCode } from '@components/notion/NCode';
-export { default as NNumberedListItem } from '@components/notion/NNumberedListItem';
-export { default as NParagraph } from '@components/notion/NParagraph';
-export { default as NQuote } from '@components/notion/NQuote';
-export { default as NTodo } from '@components/notion/NTodo';
-export { default as NToggle } from '@components/notion/NToggle';
+export { default as NHeading1 } from '@components/notion/heading1';
+export { default as NHeading2 } from '@components/notion/heading2';
+export { default as NHeading3 } from '@components/notion/heading3';
+export { default as NBookmark } from '@components/notion/bookmark';
+export { default as NBulletedList } from '@components/notion/bulletedList';
+export { default as NCallout } from '@components/notion/callout';
+export { default as NCode } from '@components/notion/code';
+export { default as NNumberedListItem } from '@components/notion/numberedList';
+export { default as NParagraph } from '@components/notion/paragraph';
+export { default as NQuote } from '@components/notion/quote';
+export { default as NTodo } from '@components/notion/todo';
+export { default as NToggle } from '@components/notion/toggle';
