+++ src/components/notion/numberedList/index.scss
@@ -1,0 +1,4 @@
+.block-numbered-list {
+  line-height: 32px;
+  padding-left: 20px;
+}
