+++ src/components/notion/numberedList/index.tsx
@@ -1,19 +1,18 @@
 import * as React from 'react';
-import '@scss/notion.scss';
+import './index.scss';
 import { NParagraph } from '@components/notion';
 import { Children } from '@types';
 import ContentWrapper from '@components/post/contents';
 
-interface NBulletedListProps {
-  bulletedListItem: Children;
+interface NNumberedListProps {
+  numberedListItem: Children;
 }
 
-export default function NBulletedList({ bulletedListItem }: NBulletedListProps) {
-  const { has_children, children } = bulletedListItem;
-
+export default function NNumberedList({ numberedListItem }: NNumberedListProps) {
+  const { has_children, children } = numberedListItem;
   return (
     <>
-      <NParagraph paragraph={bulletedListItem.bulleted_list_item} />
+      <NParagraph paragraph={numberedListItem.numbered_list_item} />
       {has_children && children?.length > 0 && <ContentWrapper childrens={children} />}
     </>
   );
