+++ src/components/notion/quote/index.scss
@@ -1,0 +1,8 @@
+.block-quote {
+  line-height: 1.75rem;
+  margin: 12px 0px;
+  padding: 8px 12px;
+  border-left: 4px solid var(--color-primary);
+  width: 100%;
+  box-sizing: border-box;
+}
