+++ src/components/notion/todo/index.scss
@@ -1,0 +1,10 @@
+.block-todo {
+  line-height: 1.75rem;
+  display: flex;
+  align-items: center;
+
+  input[type='checkbox'] {
+    margin-left: 8px;
+    margin-right: 4px;
+  }
+}
