+++ src/components/notion/toggle/index.scss
@@ -1,0 +1,23 @@
+.block-toggle {
+  line-height: 1.5rem;
+  margin: 8px 0px 8px 4px;
+
+  .toggle-title-box {
+    display: flex;
+    align-items: center;
+    cursor: pointer;
+
+    .icon-box {
+      margin-top: 2px;
+      margin-right: 6px;
+      transform: rotate(0);
+      transition-duration: 0.1s;
+      &.open {
+        transform: rotate(90deg);
+      }
+    }
+  }
+  .toggle-content-box {
+    margin-left: 20px;
+  }
+}
