+++ src/scss/notion.scss
@@ -1,157 +1,1 @@
-@import '@scss/abstracts/mixin';
 
-.block-paragraph {
-  line-height: 1.75rem;
-  width: 100%;
-
-  code {
-    background-color: var(--color-grey-6);
-    color: var(--color-red);
-    padding: 2px 6px;
-    border-radius: 4px;
-    font-size: 11pt;
-  }
-}
-
-.block-paragraph-text {
-  display: inline;
-  line-height: 1.75rem;
-  width: fit-content;
-}
-
-.block-bulleted-list {
-  line-height: 32px;
-  padding-left: 20px;
-
-  ::marker {
-    unicode-bidi: isolate;
-    font-variant-numeric: tabular-nums;
-    text-transform: none;
-    text-indent: 0px !important;
-    text-align: start !important;
-    text-align-last: start !important;
-  }
-}
-
-.block-numbered-list {
-  line-height: 32px;
-  padding-left: 20px;
-}
-
-.block-quote {
-  line-height: 1.75rem;
-  margin: 12px 0px;
-  padding: 8px 12px;
-  border-left: 4px solid var(--color-primary);
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.block-todo {
-  line-height: 1.75rem;
-  display: flex;
-  align-items: center;
-
-  input[type='checkbox'] {
-    margin-left: 8px;
-    margin-right: 4px;
-  }
-}
-
-.block-callout {
-  line-height: 1.75rem;
-  display: flex;
-  flex-direction: column;
-  background-color: var(--color-shadow);
-  box-shadow: 0px 6px 5px -4px var(--color-grey-3);
-  padding: 12px 16px 12px 4px;
-  padding-left: 40px;
-  margin: 24px 0px;
-  border-radius: 8px;
-  position: relative;
-  color: var(--color-base);
-  .icon-box {
-    position: absolute;
-    top: 10px;
-    left: 10px;
-    font-size: 1.5rem;
-  }
-}
-
-.block-toggle {
-  line-height: 1.5rem;
-  margin: 8px 0px 8px 4px;
-
-  .toggle-title-box {
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-
-    .icon-box {
-      margin-top: 2px;
-      margin-right: 6px;
-      transform: rotate(0);
-      transition-duration: 0.1s;
-      &.open {
-        transform: rotate(90deg);
-      }
-    }
-  }
-  .toggle-content-box {
-    margin-left: 20px;
-  }
-}
-
-.block-code {
-  line-height: 1.75rem;
-  background-color: #efefef;
-  margin: 16px 0px;
-  border-radius: 4px;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-
-  pre {
-    white-space: pre-wrap;
-    width: 100%;
-    margin: 0;
-    border-radius: 4px;
-    box-sizing: border-box;
-  }
-
-  span.linenumber {
-    user-select: none;
-    -webkit-user-select: none;
-  }
-
-  .code-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 8px 12px 0px;
-
-    .language {
-      color: var(--color-grey);
-      font-size: 9pt;
-    }
-  }
-
-  .caption {
-    color: var(--color-grey);
-    padding: 2px 12px;
-    font-size: 9pt;
-  }
-}
-
-.block-image {
-  width: 100%;
-  margin: 0;
-  img {
-    width: 100%;
-    border-radius: 4px;
-  }
-  .caption-box {
-    color: var(--color-grey);
-    font-size: 10pt;
-  }
-}
